Rename iOS TestFlight test filenames to avoid build failure in automation
